moved to update

+++ acceptance/acceptance_test.go
@@ -270,11 +270,13 @@ func testAccept(t *testing.T, inprocessMode bool, singleTest string) int {
 			config, configPath := internal.LoadConfig(t, dir)
 			skipReason := getSkipReason(&config, configPath)
 
-			// Generate materialized config for this test
-			// We do this before skipping the test, so the configs are generated for all tests.
-			materializedConfig, err := internal.GenerateMaterializedConfig(config)
-			require.NoError(t, err)
-			testutil.WriteFile(t, filepath.Join(dir, internal.MaterializedConfigFile), materializedConfig)
+			if testdiff.OverwriteMode {
+				// Generate materialized config for this test
+				// We do this before skipping the test, so the configs are generated for all tests.
+				materializedConfig, err := internal.GenerateMaterializedConfig(config)
+				require.NoError(t, err)
+				testutil.WriteFile(t, filepath.Join(dir, internal.MaterializedConfigFile), materializedConfig)
+			}
 
 			if skipReason != "" {
 				skippedDirs += 1
@@ -903,7 +905,8 @@ func CopyDir(src, dst string, inputs, outputs map[string]bool) error {
 			return err
 		}
 
-		if strings.HasPrefix(relPath, "out") {
+		_, isInput := inputs[relPath]
+		if strings.HasPrefix(relPath, "out") && !isInput {
 			if !info.IsDir() {
 				outputs[relPath] = true
 			}
